box: factor out definition of `mh_strnu32_t` from `tuple_dictionary.c`
Hash tables mapping strings to integers often occur: factor out definition
of `mh_strnu32_t` —  it will be reused in scope of #7897.

Add `mh_strnu32_find_str`: a helper for looking up a key in the
`mh_strnu32_t` hash table, similar to `mh_strnptr_find_str`.

Needed for #7897

NO_CHANGELOG=refactoring
NO_DOC=refactoring
NO_TEST=refactoring
